services/fraud-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.4.0
python-dotenv==1.0.1
pydantic==2.6.4
scikit-learn==1.4.1
numpy==1.26.4
